+++ packages/core/src/plugins/index.ts
@@ -3,3 +3,4 @@ export * from "./executor";
 export * from "./registry";
 export * from "./type-resolver";
 export * from "./schema-helper";
+export * from "./create";